pyproject.toml
[project]
name = "pacorain-monorepo"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "click>=8.0.0,<9.0.0",
    "pulumi>=3.0.0,<4.0.0",
    "pulumi-aws>=7.0.0,<8.0.0",
    "pulumi-proxmoxve>=7.11.0,<8.0.0",
    "PyYAML>=6.0,<7.0",
]

[tool.setuptools.packages.find]
where = ["src/python"]

[project.scripts]
lab = "lab_cli.__main__:main"

[tool.uv]
package = true




